# PERF: statically define classes for is_dtype checks