@jreback 2020-04-07
i’m fact could simply instantiate classes