@jbrockmendel 2020-04-07
Nice, I didnt expect `classes` to account for that much overhead.