@WillAyd 2020-04-07
> I can't run asv right now ([#33315 (comment)](https://github.com/pandas-dev/pandas/issues/33315#issuecomment-610329344)) but here are some timeits

Hmm don't think that issue is the root of any ASV issues - are you getting any other error messages from the run? Maybe worth opening a dedicated issue