@TomAugspurger 2020-04-07
Which logic would you move? It seems like `_is_dtype_or_type` requires a callable, and I want to keep the definition of that callable out of e.g. `is_bool_dtype`.