@TomAugspurger 2020-04-16
Probably a non-issue (or less pressing) after #33400 is done. I'll re-run
benchmarks after that's merged.

On Thu, Apr 16, 2020 at 12:23 PM jbrockmendel <notifications@github.com>
wrote:

> @TomAugspurger <https://github.com/TomAugspurger> is this still something
> you want to merge, or is it superceded by #33400
> <https://github.com/pandas-dev/pandas/pull/33400>?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/pandas-dev/pandas/pull/33364#issuecomment-614786585>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAKAOISFT56BCX3OG7EYTJ3RM45J5ANCNFSM4MDBB6BQ>
> .
>